requirements.txt
droidrun
openai
llama-index
pydantic
python-dotenv
